Use broccoli-caching-writer.
Even though libsass is insanely fast, we still see broccoli-sass in the
"slow-trees listing" (at around 250ms) for every rebuild regardless of
what files were changed.

This change updates to inherit from broccoli-caching-writer, which
simply symlinks (using symlink-or-copy package) from the last runs
cache directory if the input trees have not changed.

The result of this is that when changing files in a tree other than those
provided to broccoli-sass the broccoli-sass tree rebuilds are nearly
instant ( less than 10ms on SSD ).